Allow any minor version of Rails 5
This gem doesn't currently have a need to prevent anyone from
using Rails up to 5.2.x, so this change allows that. It's
possible we could support Rails 6, but that hasn't been tested.

This commit also fixes a reference to the test Rakefile, which
is in a different folder than originally specified.