Fix tasks and commands so they can be run from the root directory
fixes lessonly/scim_rails#25
fixes lessonly/scim_rails#26

Why
Since we are using the spec/ director instead of the default test/ we
need to specify the locations of the dummy application and it's rake
file. With out this, you cannot run things like db:migrate from the root
of the engine as you should. This goes for any rake task you would
usually call from the root. Additionally, commands like rails console
would not work because it would not be able to find the dummy
application to load it.