Remove Gemfile.lock and update Ruby
The Gemfile.lock is not a file that we want checked in with the gem. The
purpose of a Gemfile.lock is to lock in a combination of dependencies
for an application. In the case of a gem, we are just specifying the
limits for our dependency versions in the gemspec and are not locking in
a combination of dependencies (since ultimately, the application is
in charge of it's own dependencies).

The only effect the Gemfile.lock has in this project is for when running
it in the dummy app; is has no effect on any of the applications it's
installed in. The gem should work against the latest specified in it's
gemspec so there is no reason to pin the dummy app to the version in a
Gemfile.lock checked into source control.

More info in this blog post: https://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/