Fix patch endpoint archiving and case sensitivity
fixes #32

After a fix shipped in #29, you could no longer de-provision a user
because when false was passed in as the value for active, it thought
there was no value for active.

fixes #33

Technically according to the specifications, the value for op is only
valid if it is one of three lower case values. However, our previous
implementation accepted a case insensitive version. Because of this, we
are adding backward compatibility for case insensitivity until we have a
chance to properly make a breaking change and only allow lower case.